# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Contratos",
        "description": "API para gerenciamento de contratos, convênios, serviços e bancos.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/contratos": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Listar contratos",
                "description": "Retorna uma lista de contratos.",
                "operationId": "11f09a10344e23955021d6249957653e",
                "responses": {
                    "200": {
                        "description": "Lista de contratos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nome_banco": {
                                                "type": "string",
                                                "example": "Banco A"
                                            },
                                            "verba": {
                                                "type": "number",
                                                "example": 10000
                                            },
                                            "codigo_contrato": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "data_inclusao": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2023-01-01"
                                            },
                                            "valor": {
                                                "type": "number",
                                                "example": 5000
                                            },
                                            "prazo": {
                                                "type": "integer",
                                                "example": 12
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contratos não encontrados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Criar um contrato",
                "description": "Cria um novo contrato.",
                "operationId": "8baa1a8b67f0d0ff7c3e54bc36bb7f1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contrato"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contrato criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "codigo": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "prazo": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "valor": {
                                            "type": "number",
                                            "example": 5000
                                        },
                                        "data_inclusao": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-01-01"
                                        },
                                        "convenio_servico": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contrato": {
                "title": "Contrato",
                "required": [
                    "prazo",
                    "valor",
                    "data_inclusao",
                    "convenio_servico"
                ],
                "properties": {
                    "prazo": {
                        "description": "Prazo do contrato em meses",
                        "type": "integer",
                        "example": 12
                    },
                    "valor": {
                        "description": "Valor do contrato",
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "data_inclusao": {
                        "description": "Data de inclusão do contrato",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "convenio_servico": {
                        "description": "Código do convênio serviço",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Contratos",
            "description": "Contratos"
        }
    ]
}